feat: use BigInteger instead of Long

+++ product-service/src/main/java/com/aanbari/productservice/dto/ProductDto.java
@@ -3,6 +3,8 @@ package com.aanbari.productservice.dto;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigInteger;
+
 @Builder
 @Getter
 public class ProductDto {
@@ -10,5 +12,5 @@ public class ProductDto {
     private String name;
     private String tag;
     private String description;
-    private Long price;
+    private BigInteger price;
 }
